# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 用户管理服务
  description: 基于 FastAPI、Pydantic、SQLAlchemy 与 Redis 的用户、角色和视频分析任务服务。
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: 本地开发环境
...
paths:
...
  /api/video/upload:
    post:
      tags: [视频分析任务]
      summary: 上传视频分析任务
      description: 创建任务后立即返回任务 ID，并在后台异步执行视频分析。单设备最多同时存在 2 个活跃任务。
      operationId: upload_video
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoTaskCreate"
            example:
              deviceId: device_001
              videoUrl: https://example.com/videos/demo.mp4
              duration: 30
      responses:
        "200":
          description: 创建结果；并发超限时 code=429，Redis 不可用时 code=50002
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoTaskUploadApiResponse"
  /api/video/status/{taskId}:
    get:
      tags: [视频分析任务]
      summary: 查询视频分析任务状态
      operationId: get_video_status
      parameters:
        - name: taskId
          in: path
          required: true
          description: 32 位小写十六进制任务 ID
          schema:
            type: string
            minLength: 32
            maxLength: 32
            pattern: '^[0-9a-f]{32}$'
      responses:
        "200":
          description: 任务状态；任务不存在时 code=40400
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoTaskApiResponse"
...
components:
...
  schemas:
...
    VideoTaskStatus:
      type: string
      description: 视频分析任务状态
      enum: [processing, success, failed]
    VideoTaskCreate:
      type: object
      required: [deviceId, videoUrl, duration]
      properties:
        deviceId:
          type: string
          minLength: 1
          maxLength: 32
          description: 设备 ID
        videoUrl:
          type: string
          format: uri
          maxLength: 2048
          description: HTTP/HTTPS 视频地址
        duration:
          type: integer
          minimum: 1
          maximum: 86400
          description: 视频时长（秒）
    VideoTaskUploadOut:
      type: object
      required: [taskId]
      properties:
        taskId:
          type: string
          minLength: 32
          maxLength: 32
          pattern: '^[0-9a-f]{32}$'
          description: 视频分析任务 ID
    VideoTaskOut:
      type: object
      required: [taskId, status, resultUrl, createTime]
      properties:
        taskId:
          type: string
          minLength: 32
          maxLength: 32
          pattern: '^[0-9a-f]{32}$'
          description: 视频分析任务 ID
        status:
          $ref: "#/components/schemas/VideoTaskStatus"
        resultUrl:
          type: string
          format: uri
          nullable: true
          description: 分析结果地址，任务成功前为空
        createTime:
          type: string
          format: date-time
          description: 任务创建时间
...
    VideoTaskUploadApiResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          description: 0 成功；429 单设备活跃任务超限；42200 参数校验失败；50001 创建失败；50002 Redis 不可用
        message:
          type: string
        data:
          allOf:
            - $ref: "#/components/schemas/VideoTaskUploadOut"
          nullable: true
    VideoTaskApiResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          description: 0 成功；40400 任务不存在；42200 参数校验失败
        message:
          type: string
        data:
          allOf:
            - $ref: "#/components/schemas/VideoTaskOut"
          nullable: true
...